moved .combine method to stitching.Dataset class as a staticmethod
